restructuring of surf_react parent class
