+++ src/surf_react_global.h
@@ -30,7 +30,6 @@ class SurfReactGlobal : public SurfReact {
   SurfReactGlobal(class SPARTA *, int, char **);
   ~SurfReactGlobal();
   int react(Particle::OnePart *&, int, double *, Particle::OnePart *&);
-
   char *reactionID(int);
   int match_reactant(char *, int) {return 1;}
   int match_product(char *, int) {return 1;}
