+++ src/surf_react_prob.h
@@ -31,9 +31,44 @@ class SurfReactProb : public SurfReact {
   ~SurfReactProb();
   void init();
   int react(Particle::OnePart *&, int, double *, Particle::OnePart *&);
+  char *reactionID(int);
+  int match_reactant(char *, int);
+  int match_product(char *, int);
 
  private:
   class RanPark *random;     // RNG for reaction probabilities
+
+  // reaction info, as read from file
+
+  struct OneReaction {
+    int active;                    // 1 if reaction is active
+    int type;                      // reaction type = DISSOCIATION, etc
+    int style;                     // reaction style = ARRHENIUS, etc
+    int ncoeff;                    // # of numerical coeffs
+    int nreactant,nproduct;        // # of reactants and products
+    char **id_reactants,**id_products;  // species IDs of reactants/products
+    int *reactants,*products;      // species indices of reactants/products
+    double *coeff;                 // numerical coeffs for reaction
+    char *id;                      // reaction ID (formula)
+  };
+
+  OneReaction *rlist;              // list of all reactions read from file
+  int nlist_prob;                  // # of reactions read from file
+  int maxlist_prob;                // max # of reactions in rlist
+
+  // possible reactions a reactant species is part of
+
+  struct ReactionI {
+    int *list;           // list of indices into rlist, ptr into indices
+    int n;               // # of reactions in list
+  };
+
+  ReactionI *reactions;       // reactions for all species
+  int *indices;               // master list of indices
+
+  void init_reactions();
+  void readfile(char *);
+  int readone(char *, char *, int &, int &);
 };
 
 }
